VIM:  Enable ^N completion on terms starting with @
Lots of crazy hacking around to make it so that vim completion can work
on words that start with '@'.  The vim-pencil plugin totally clobbers
iskeyword() on a buffer by buffer basis.  Wrote a function to add '@-@'
to iskeyword on a buffer-local basis and made sure it was called after
pencil:init().